[Lab 08a] Ex - Refactored Package structure for SpotMapView/Presenter and deleted empty activities folder
