+++ app/src/main/java/com/archaeologicalfieldwork/views/map/SpotMapPresenter.kt
@@ -1,4 +1,4 @@
-package com.archaeologicalfieldwork.activities
+package com.archaeologicalfieldwork.views.map
 
 
 import com.google.android.gms.maps.CameraUpdateFactory
